Change properties storage
add test files

Changed to RNA properties to store object list and other value in the blender file

updated light creation and main parent creation

Camera and animation need rework

Created Values properties

+++ SyntheticRTI-Plugin/SyntheticRTI.py
@@ -2,10 +2,9 @@ import bpy
 from mathutils import Vector
 
 ###Global var###
-
-srti_lamp_list = []
-srti_camera_list = []
-srti_main_parent = []
+#srti_lamp_list = []
+#srti_camera_list = []
+#srti_main_parent = []
 
 ###Various function###
 
@@ -18,52 +17,68 @@ def printdebug():
     print("main:")
     print(srti_main_parent)
 
-def populate(context):
-    """populate global variables from the main parent name when loading a file"""
-    global srti_main_parent 
-    global srti_lamp_list
-    global srti_camera_list
+    #deprecated
+    #def populate(context):
+    #    """populate global variables from the main parent name when loading a
+    #    file"""
+    #    global srti_main_parent
+    #    global srti_lamp_list
+    #    global srti_camera_list
 
-    parent_name = context.scene.srti_main_parent_prop
-    if parent_name:
-        print(parent_name)
-        main_parent = context.scene.objects[parent_name]
-        srti_main_parent.append(main_parent)
+    #    parent_name = context.scene.srti_main_parent_prop
+    #    if parent_name:
+    #        print(parent_name)
+    #        main_parent = context.scene.objects[parent_name]
+    #        srti_main_parent.append(main_parent)
         
-        for obj in context.scene.objects:
-            if obj.parent == main_parent:
-                if obj.type == 'LAMP': #take all lights
-                    srti_lamp_list.append(obj)
-                elif obj.type == 'CAMERA': #take all cameras
-                    srti_camera_list.append(obj)
+    #        for obj in context.scene.objects:
+    #            if obj.parent == main_parent:
+    #                if obj.type == 'LAMP': #take all lights
+    #                    srti_lamp_list.append(obj)
+    #                elif obj.type == 'CAMERA': #take all cameras
+    #                    srti_camera_list.append(obj)
 
-        printdebug()
-    else:
-        print("Nothing to load")
+    #        printdebug()
+    #    else:
+    #        print("Nothing to load")
+
+#def update_objects_lists():
+#    """If there are no scene idproperties it creates empty ones for every
+#    scene"""
+#    for scene in bpy.data.scenes:
+#        try:
+#            scene['srti_lamp_list']
+#        except:
+#            scene['srti_lamp_list'] = []
+
+#        try:
+#            scene['srti_main_parent']
+#        except:
+#            scene['srti_main_parent'] = []
+
+#        try:
+#            scene['srti_camera_list']
+#        except:
+#            scene['srti_camera_list'] = []
 
 def create_main(scene):
     """Create the main object if not already present"""
     #Aggiungo un empty a cui collegare tutte le luci per poter modificarle come gruppo
-    global srti_main_parent 
-
-    if not srti_main_parent:
+ 
+    if not scene.srti_props.main_parent:
         main_parent = bpy.data.objects.new(name = "main_parent", object_data = None)
         scene.objects.link(main_parent)
         main_parent.empty_draw_type = 'SPHERE'
-   
-        srti_main_parent.append(main_parent) 
-        scene.srti_main_parent_prop = main_parent.name
+        scene.srti_props.main_parent = main_parent
 
 def delete_main(scene):
-    """Delete main parent"""
-    global srti_main_parent 
-
-    bpy.ops.object.select_all(action='DESELECT')
-    if srti_main_parent:
-        srti_main_parent[0].select = True #delete the main parent (need revision)
-    bpy.ops.object.delete()
-    srti_main_parent = []
-    scene.srti_main_parent_prop = "" #delete the property of the main parent
+    """Delete main parent""" 
+    if scene.srti_props.main_parent:
+        bpy.ops.object.select_all(action='DESELECT')
+        scene.srti_props.main_parent.select = True #delete the main parent (need revision)
+        scene.srti_props.main_parent = None
+        bpy.ops.object.delete()
+   
 
 ###OPERATORS###
 #########LAMP#########
@@ -77,11 +92,11 @@ class create_lamps(bpy.types.Operator):
         bpy.ops.srti.delete_lamps() #delete exsisting lamps
 
         curr_scene = context.scene
-        file_path = curr_scene.srti_lightsFilePath
+        file_path = curr_scene.srti_props.lightsFilePath
 
         #create the main parent
         create_main(curr_scene)
-        main_parent = srti_main_parent[0]
+        main_parent = curr_scene.srti_props.main_parent
 
         file = open(file_path)
         righe = file.readlines() #copio tutte le linee in memoria
@@ -92,10 +107,9 @@ class create_lamps(bpy.types.Operator):
         print(n_luci)
         
         ##Uso una luce standard, TODO aggiungere un oggetto
-        lamp_data = bpy.data.lamps.new(name="Luce_progetto", type='SUN') # Create new lamp datablock. It s going to be created outside
-
+        lamp_data = bpy.data.lamps.new(name="Luce_progetto", type='SUN') # Create new lamp datablock.  It s going to be created outside
         
-        for lamp in range(1, n_luci + 1): #passo tutte le luci 
+        for lamp in range(1, n_luci + 1): #passo tutte le luci
             valori_riga = righe[lamp].split() #divido i valori
             lmp_x = float(valori_riga[1])
             lmp_y = float(valori_riga[2])
@@ -111,8 +125,9 @@ class create_lamps(bpy.types.Operator):
             lamp_object.rotation_mode = 'QUATERNION'
             lamp_object.rotation_quaternion = direction.to_track_quat('Z','Y')
             
-            lamp_object.keyframe_insert(data_path = 'hide_render', frame = i-1)
-            srti_lamp_list.append(lamp_object)
+            ##change the name
+            lamp = curr_scene.srti_list_lights.add()
+            lamp.light = lamp_object
 
         return{'FINISHED'}
 
@@ -123,21 +138,20 @@ class delete_lamps(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     def execute(self, context):
-        global srti_lamp_list
+        lamp_list = context.scene.srti_list_lights
 
         bpy.ops.object.select_all(action='DESELECT')
 
-        if srti_lamp_list:
-            for obj in srti_lamp_list:
-                obj.select = True
+        if lamp_list:
+            for lamp_obj in lamp_list:
+                lamp_obj.light.select = True
     
         bpy.ops.object.delete()
 
-        if not srti_camera_list: #delete the main object if no cameras
+        if not context.scene.srti_list_cameras: #delete the main object if no cameras
             delete_main(context.scene)
-
-        srti_lamp_list = [] #delete information in global variable
-
+        
+        context.scene.srti_list_lights.clear() #delete the idlist
         return{'FINISHED'}
 
 ########CAMERA########
@@ -154,7 +168,7 @@ class create_cameras(bpy.types.Operator):
 
         #create the main parent
         create_main(curr_scene)
-        main_parent = srti_main_parent[0]
+        main_parent = curr_scene['srti_main_parent']
 
         camera_data = bpy.data.cameras.new("Camera")
         camera_object = bpy.data.objects.new("Camera", camera_data)
@@ -163,7 +177,12 @@ class create_cameras(bpy.types.Operator):
         camera_object.parent = main_parent
         camera_object.location = (0, 0, 2)
 
-        srti_camera_list.append(camera_object)
+        if not curr_scene['srti_camera_list']:
+            curr_scene['srti_camera_list'] = [camera_object]
+            print("una camera")
+        else:
+            curr_scene['srti_camera_list'].extend([camera_object])
+            print("più camere")
 
         return{'FINISHED'}
 
@@ -174,19 +193,22 @@ class delete_cameras(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO'}
     
     def execute(self, context):
-        global srti_camera_list
+        curr_scene = context.scene
+
+        camera_list = curr_scene['srti_camera_list']
+
         bpy.ops.object.select_all(action='DESELECT')
         
-        if srti_camera_list:
-            for obj in srti_camera_list:
+        if camera_list:
+            for obj in camera_list:
                 obj.select = True
     
         bpy.ops.object.delete()
 
-        if not srti_lamp_list: #delete the main object if no lamps
+        if not curr_scene['srti_lamp_list']: #delete the main object if no lamps
              delete_main(context.scene)
         
-        srti_camera_list = []
+        curr_scene['srti_camera_list'] = []
               
         return{'FINISHED'}
 
@@ -236,8 +258,8 @@ class animate_all(bpy.types.Operator):
                 curr_frame = (index_prop * tot_cam * tot_light) + (index_cam * tot_light) + index_light + 1
                 #hide lamp on theprevious and next frame
                 lamp.hide_render = True
-                lamp.keyframe_insert(data_path = 'hide_render', frame = curr_frame-1)
-                lamp.keyframe_insert(data_path = 'hide_render', frame = curr_frame+1)
+                lamp.keyframe_insert(data_path = 'hide_render', frame = curr_frame - 1)
+                lamp.keyframe_insert(data_path = 'hide_render', frame = curr_frame + 1)
                 #rendo visibile la lampada solo nel suo frame
                 lamp.hide_render = False
                 lamp.keyframe_insert(data_path = 'hide_render', frame = curr_frame)
@@ -273,7 +295,6 @@ class create_export_file(bpy.types.Operator):
 ##############
 #####GUI######
 ##############
-
 class SyntheticRTIPanel(bpy.types.Panel):
     """Creates a Panel in the Object properties window"""
     bl_label = "Lamps"
@@ -283,31 +304,70 @@ class SyntheticRTIPanel(bpy.types.Panel):
     bl_category = "SyntheticRTI"
 
     def draw(self, context):
+        curr_scene = context.scene
+
+        num_light = len(curr_scene.srti_list_lights)
+        num_cam = len(curr_scene.srti_list_cameras)
+
         layout = self.layout
-        ##layout.operator("")
-        layout.prop(context.scene, "srti_lightsFilePath", text = 'File')
+        layout.prop(curr_scene.srti_props, "lightsFilePath", text = 'File')
         layout.operator("srti.create_lamps")
         layout.operator("srti.delete_lamps")
         layout.operator("srti.create_cameras")
         layout.operator("srti.delete_cameras")
-        layout.label("main = %s" % (str(len(srti_main_parent))))
-        layout.label("lamps = %s" % (str(len(srti_lamp_list))))
-        layout.label("cameras = %s" % (str(len(srti_camera_list))))
-        layout.label("frame totali = %s" % (str(len(srti_camera_list) * len(srti_lamp_list))))
+        box = layout.box()
+        box.label("RNA PROP")
+        box.label("main = %s" % curr_scene.srti_props.main_parent)
+        box.label("lamps = %i" % num_light)
+        box.label("cameras = %i" % num_cam)
+        box.label("frame totali = %i" % (num_light * num_cam))
         layout.operator("srti.animate_all")
         layout.operator("srti.render_images")
         layout.operator("srti.create_file")
 
+
+class srti_props(bpy.types.PropertyGroup):
+    lightsFilePath = bpy.props.StringProperty(name="Lights file Path",
+        subtype='FILE_PATH',
+        default="*.lp",
+        description = 'Path to the lights file.')
+
+    main_parent = bpy.props.PointerProperty(name="Main Parent",
+        type=bpy.types.ID,
+        description = "Main object of the group")
+
+class light(bpy.types.PropertyGroup):
+    light = bpy.props.PointerProperty(name="Light object",
+        type = bpy.types.Object,
+        description = "A lamp")
+
+class camera(bpy.types.PropertyGroup):
+    camera = bpy.props.PointerProperty(name = "Camera object",
+        type = bpy.types.Object,
+        description = "A camera")
+
+class value(bpy.types.PropertyGroup):
+    name = bpy.props.StringProperty()
+    min = bpy.props.FloatProperty()
+    max = bpy.props.FloatProperty()
+    steps = bpy.props.IntProperty()
+
+
 def register():
     ##Scene data definition
-    ##File path of selected lights file
-    bpy.types.Scene.srti_lightsFilePath = bpy.props.StringProperty(
-            name="Lights file Path",
-            subtype='FILE_PATH',
-            default="*.lp",
-            description = 'Path to the lights file.'      
-            )
-    bpy.types.Scene.srti_main_parent_prop = bpy.props.StringProperty()
+    bpy.utils.register_class(srti_props)
+    bpy.types.Scene.srti_props = bpy.props.PointerProperty(type = srti_props)
+
+    bpy.utils.register_class(light)
+    bpy.types.Scene.srti_list_lights = bpy.props.CollectionProperty(type = light)
+
+    bpy.utils.register_class(camera)
+    bpy.types.Scene.srti_list_cameras = bpy.props.CollectionProperty(type = camera)
+
+    bpy.utils.register_class(value)
+    bpy.types.Scene.srti_list_values = bpy.props.CollectionProperty(type = value)
+
+    ##resgister properties
     bpy.utils.register_class(delete_lamps)
     bpy.utils.register_class(delete_cameras)
     bpy.utils.register_class(create_lamps)
@@ -316,8 +376,8 @@ def register():
     bpy.utils.register_class(render_images)
     bpy.utils.register_class(animate_all)
     bpy.utils.register_class(SyntheticRTIPanel)
-    populate(bpy.context)
-
+    #update_objects_lists()
+    #populate(bpy.context)
 def unregister():
     bpy.utils.unregister_class(SyntheticRTIPanel)
     bpy.utils.unregister_class(create_cameras)
